planning/requirements.txt
Flask
google-api-python-client
oauth2client
python-dotenv
pika
flask_sqlalchemy
